Move Generators to a file to have cleaner code
